CardView und ScoreboardView ausgelagert wegen duplications
