+++ .idea/shelf/Uncommitted_changes_before_Checkout_at_23_06_2025_10_56_[Changes]/shelved.patch
@@ -1,0 +1,41 @@
+Index: .idea/deploymentTargetSelector.xml
+IDEA additional info:
+Subsystem: com.intellij.openapi.diff.impl.patch.BaseRevisionTextPatchEP
+<+><?xml version=\"1.0\" encoding=\"UTF-8\"?>\r\n<project version=\"4\">\r\n  <component name=\"deploymentTargetSelector\">\r\n    <selectionStates>\r\n      <SelectionState runConfigName=\"app\">\r\n        <option name=\"selectionMode\" value=\"DROPDOWN\" />\r\n        <DropdownSelection timestamp=\"2025-06-15T08:13:23.617459200Z\">\r\n          <Target type=\"DEFAULT_BOOT\">\r\n            <handle>\r\n              <DeviceId pluginId=\"LocalEmulator\" identifier=\"path=C:\\Users\\HP\\.android\\avd\\Pixel_6.avd\" />\r\n            </handle>\r\n          </Target>\r\n        </DropdownSelection>\r\n        <DialogSelection />\r\n      </SelectionState>\r\n    </selectionStates>\r\n  </component>\r\n</project>
+Subsystem: com.intellij.openapi.diff.impl.patch.CharsetEP
+<+>UTF-8
+===================================================================
+diff --git a/.idea/deploymentTargetSelector.xml b/.idea/deploymentTargetSelector.xml
+--- a/.idea/deploymentTargetSelector.xml	(revision 121f8a648dfe2595e15253b7298e45b3a3591165)
++++ b/.idea/deploymentTargetSelector.xml	(date 1750640626500)
+@@ -3,7 +3,7 @@
+   <component name="deploymentTargetSelector">
+     <selectionStates>
+       <SelectionState runConfigName="app">
+-        <option name="selectionMode" value="DROPDOWN" />
++        <option name="selectionMode" value="DIALOG" />
+         <DropdownSelection timestamp="2025-06-15T08:13:23.617459200Z">
+           <Target type="DEFAULT_BOOT">
+             <handle>
+@@ -11,7 +11,20 @@
+             </handle>
+           </Target>
+         </DropdownSelection>
+-        <DialogSelection />
++        <DialogSelection>
++          <targets>
++            <Target type="DEFAULT_BOOT">
++              <handle>
++                <DeviceId pluginId="LocalEmulator" identifier="path=C:\Users\HP\.android\avd\Pixel_5.avd" />
++              </handle>
++            </Target>
++            <Target type="DEFAULT_BOOT">
++              <handle>
++                <DeviceId pluginId="LocalEmulator" identifier="path=C:\Users\HP\.android\avd\Pixel_6.avd" />
++              </handle>
++            </Target>
++          </targets>
++        </DialogSelection>
+       </SelectionState>
+     </selectionStates>
+   </component>
