+++ .idea/shelf/Uncommitted_changes_before_Checkout_at_23_06_2025_10_56__Changes_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Checkout_at_23_06_2025_10_56_[Changes]" date="1750668969737" recycled="true" deleted="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_23_06_2025_10_56_[Changes]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 23.06.2025 10:56 [Changes]" />
+</changelist>
