+++ .idea/shelf/Uncommitted_changes_before_Checkout_at_23_06_2025_15_52_[Changes]/shelved.patch
@@ -1,0 +1,41 @@
+Index: .idea/deploymentTargetSelector.xml
+IDEA additional info:
+Subsystem: com.intellij.openapi.diff.impl.patch.BaseRevisionTextPatchEP
+<+><?xml version=\"1.0\" encoding=\"UTF-8\"?>\r\n<project version=\"4\">\r\n  <component name=\"deploymentTargetSelector\">\r\n    <selectionStates>\r\n      <SelectionState runConfigName=\"app\">\r\n        <option name=\"selectionMode\" value=\"DROPDOWN\" />\r\n        <DropdownSelection timestamp=\"2025-06-16T20:38:59.388440Z\">\r\n          <Target type=\"DEFAULT_BOOT\">\r\n            <handle>\r\n              <DeviceId pluginId=\"LocalEmulator\" identifier=\"path=/Users/angelina/.android/avd/Pixel_3a_API_34_extension_level_7_arm64-v8a.avd\" />\r\n            </handle>\r\n          </Target>\r\n        </DropdownSelection>\r\n        <DialogSelection />\r\n      </SelectionState>\r\n    </selectionStates>\r\n  </component>\r\n</project>
+Subsystem: com.intellij.openapi.diff.impl.patch.CharsetEP
+<+>UTF-8
+===================================================================
+diff --git a/.idea/deploymentTargetSelector.xml b/.idea/deploymentTargetSelector.xml
+--- a/.idea/deploymentTargetSelector.xml	(revision 37a26a291f394a476153a22ee948371ab1239b57)
++++ b/.idea/deploymentTargetSelector.xml	(date 1750668969727)
+@@ -3,7 +3,7 @@
+   <component name="deploymentTargetSelector">
+     <selectionStates>
+       <SelectionState runConfigName="app">
+-        <option name="selectionMode" value="DROPDOWN" />
++        <option name="selectionMode" value="DIALOG" />
+         <DropdownSelection timestamp="2025-06-16T20:38:59.388440Z">
+           <Target type="DEFAULT_BOOT">
+             <handle>
+@@ -11,7 +11,20 @@
+             </handle>
+           </Target>
+         </DropdownSelection>
+-        <DialogSelection />
++        <DialogSelection>
++          <targets>
++            <Target type="DEFAULT_BOOT">
++              <handle>
++                <DeviceId pluginId="LocalEmulator" identifier="path=C:\Users\HP\.android\avd\Pixel_5.avd" />
++              </handle>
++            </Target>
++            <Target type="DEFAULT_BOOT">
++              <handle>
++                <DeviceId pluginId="LocalEmulator" identifier="path=C:\Users\HP\.android\avd\Pixel_6.avd" />
++              </handle>
++            </Target>
++          </targets>
++        </DialogSelection>
+       </SelectionState>
+     </selectionStates>
+   </component>
