+++ .idea/shelf/Uncommitted_changes_before_Checkout_at_23_06_2025_17_22__Changes_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Checkout_at_23_06_2025_17_22_[Changes]" date="1750692146790" recycled="true" deleted="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_23_06_2025_17_22_[Changes]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 23.06.2025 17:22 [Changes]" />
+</changelist>
