+++ .idea/shelf/Uncommitted_changes_before_Checkout_at_23_06_2025_19_29__Changes_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Checkout_at_23_06_2025_19_29_[Changes]" date="1750699783269" recycled="true" deleted="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_23_06_2025_19_29_[Changes]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 23.06.2025 19:29 [Changes]" />
+</changelist>
