+++ .idea/shelf/Uncommitted_changes_before_Checkout_at_23_06_2025_19_37__Changes_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Checkout_at_23_06_2025_19_37_[Changes]" date="1750700250247" recycled="false" toDelete="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_23_06_2025_19_37_[Changes]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 23.06.2025 19:37 [Changes]" />
+</changelist>
