+++ .idea/shelf/Uncommitted_changes_before_Update_at_23_06_2025_19_53__Changes_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Update_at_23_06_2025_19_53_[Changes]" date="1750701224331" recycled="true" deleted="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Update_at_23_06_2025_19_53_[Changes]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Update at 23.06.2025 19:53 [Changes]" />
+</changelist>
